casting to alleviate arm64 warnings.
Unsure if all these amends are "correct" for arm64. however, all the unit
tests still pass ... presuming the unit test coverage is robust.

+++ src/runtime/Classes/ConcreteExtensionField.m
@@ -383,7 +383,7 @@ int32_t typeSize(PBExtensionType type) {
     [output writeTag:fieldNumber format:PBWireFormatLengthDelimited];
     int32_t dataSize = 0;
     if (typeIsFixedSize(type)) {
-      dataSize = values.count * typeSize(type);
+      dataSize = (int32_t)(values.count * typeSize(type));
     } else {
       for (id value in values) {
         dataSize += [self computeSingleSerializedSizeNoTag:value];
@@ -414,7 +414,7 @@ int32_t typeSize(PBExtensionType type) {
   if (isPacked) {
     int32_t size = 0;
     if (typeIsFixedSize(type)) {
-      size = values.count * typeSize(type);
+      size = (int32_t)(values.count * typeSize(type));
     } else {
       for (id value in values) {
         size += [self computeSingleSerializedSizeNoTag:value];
