+++ src/runtime/Classes/Utilities.m
@@ -158,7 +158,7 @@ int32_t computeBoolSizeNoTag(BOOL value) {
 
 int32_t computeStringSizeNoTag(const NSString* value) {
 	const NSUInteger length = [value lengthOfBytesUsingEncoding:NSUTF8StringEncoding];
-	return computeRawVarint32Size(length) + length;
+	return computeRawVarint32Size((int32_t)length) + (int32_t)length;
 }
 
 
@@ -179,7 +179,7 @@ int32_t computeMessageSizeNoTag(const id<PBMessage> value) {
 
 
 int32_t computeDataSizeNoTag(const NSData* value) {
-	return computeRawVarint32Size(value.length) + value.length;
+	return computeRawVarint32Size((int32_t)value.length) + (int32_t)value.length;
 }
